♻️ Refactor RegisterTable usage: update imports and add editable prop; replace manual table implementation in AuxRegisters and DataRegisters with RegisterTable component
